refactor(recruit): (with-footer) 라우트 그룹 제거 및 페이지 경로 단순화 (#333)
