+++ apps/recruit/src/app/faq/page.tsx
@@ -1,6 +1,6 @@
-import {FaqSideBar} from '@/app/(with-footer)/faq/_components/FaqSideBar';
-import {FaqAccordionList} from '@/app/(with-footer)/faq/_components/FaqAccordionList';
-import {FaqContact} from '@/app/(with-footer)/faq/_components/FaqContact';
+import {FaqSideBar} from '@/app/faq/_components/FaqSideBar';
+import {FaqAccordionList} from '@/app/faq/_components/FaqAccordionList';
+import {FaqContact} from '@/app/faq/_components/FaqContact';
 import {SuspenseWrapper} from '@/components/wrappers/SuspenseWrapper';
 import Image from 'next/image';
 import HeroBanner from '@/assets/backgrounds/banners/hero-main.webp';
