+++ apps/recruit/src/app/my-page/detail/_containers/MyPageApplicationDetailContainer.tsx
