+++ apps/recruit/src/app/recruit/_components/ActivityCard.tsx
