+++ apps/recruit/src/app/recruit/_components/QualificationsCard.tsx
